+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/service/impl/SubjectCategoryServiceImpl.java
@@ -30,7 +30,7 @@ public class SubjectCategoryServiceImpl implements SubjectCategoryService {
      * @return 实例对象
      */
     @Override
-    public SubjectCategory queryById(Integer id) {
+    public SubjectCategory queryById(Long id) {
         return this.subjectCategoryDao.queryById(id);
     }
